refactor(opencode): extract session LLM request prep
